requirements_dev.txt
pip>=18.1
bump2version>=0.5.10
twine>=1.12.1
wheel>=0.32.1

#docs
sphinx-rtd-theme>=0.4.3
Sphinx>=2.2.0
sphinx-copybutton>=0.2.5
recommonmark>=0.6.0

# tests
django-app-helper>=2.0.0
selenium>=3.141.0
coverage>=4.5.1
tox>=3.13.2
mixer>=6.1.3
webdriver_manager>=1.8.2
